setup/database/requirements_comprehensive.txt
# Comprehensive Database Setup Requirements
# Core database and data processing
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.24.0

# Geospatial processing
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0
fiona>=1.9.0

# Performance and utilities
python-dotenv>=1.0.0
tqdm>=4.65.0 